pyproject.toml
[project]
name = "swo-finops-extension"
version = "0.0.0"
description = "FinOps for Cloud extension for SWO Marketplace Platform"
authors = [{ name = "SoftwareOne AG" }]
requires-python = ">=3.12,<4"
readme = "README.md"
license = "Apache-2.0"
dependencies = [
    "adaptive-cards-py>=0.3.0,<0.4",
    "aiofiles>=24.1.0,<25",
    "django==4.2.*",
    "httpx-retries>=0.4.0,<0.5",
    "jinja2==3.1.*",
    "markdown-it-py==3.0.*",
    "mpt-extension-sdk==5.3.1",
    "openpyxl==3.1.*",
    "phonenumbers==8.13.*",
    "pyairtable==2.3.*",
    "pyjwt==2.8.*",
    "pymsteams==0.2.*",
    "pytest-httpx>=0.35.0,<0.36",
    "regex>=2024.7.24,<2025",
    "requests==2.32.*",
    "types-openpyxl==3.1.*",
]

[project.scripts]
swoext = "swo.mpt.extensions.runtime.swoext:main"

[project.entry-points."swo.mpt.ext"]
app_config = "ffc.apps:ExtensionConfig"

[dependency-groups]
dev = [
    "freezegun==1.4.*",
    "ipdb==0.13.*",
    "ipython==8.*",
    "mypy>=1.15.0",
    "pre-commit==3.6.*",
    "pytest==8.2.*",
    "pytest-asyncio>=1.0.0,<2",
    "pytest-cov==4.1.*",
    "pytest-deadfixtures==2.2.*",
    "pytest-django==4.7.*",
    "pytest-httpx>=0.35.0,<0.36",
    "pytest-mock==3.12.*",
    "pytest-randomly==3.15.*",
    "pytest-xdist==3.5.*",
    "responses==0.24.*",
    "ruff==0.3.*",
    "types-aiofiles>=24.1.0.20250822,<25",
    "types-openpyxl==3.1.*",
    "types-python-dateutil>=2.9.0.20250822,<3",
    "types-requests==2.31.*",
]
sdk = [
    "django-ninja==1.1.*",
    "pyjwt==2.8.*",
    "typing-extensions==4.14.*",
]
runtime = [
    "azure-monitor-opentelemetry-exporter==1.0.0b25",
    "click==8.1.*",
    "debugpy==1.8.*",
    "gunicorn==23.0.*",
    "opentelemetry-api==1.30.*",
    "opentelemetry-instrumentation-django==0.51b0",
    "opentelemetry-instrumentation-logging==0.51b0",
    "opentelemetry-instrumentation-requests==0.51b0",
    "opentelemetry-sdk==1.30.*",
    "pyfiglet==1.0.*",
    "rich==13.7.*",
    "watchfiles==0.21.*",
]

[tool.uv]
default-groups = [
    "dev",
    "sdk",
    "runtime",
]

[tool.hatch.build.targets.sdist]
include = ["ffc"]

[tool.hatch.build.targets.wheel]
include = ["ffc"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = "tests"
pythonpath = "."
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
addopts = "--cov=ffc --cov-report=term-missing --cov-report=html --cov-report=xml"
DJANGO_SETTINGS_MODULE = "tests.django.settings"
log_cli = false
filterwarnings = [
    "ignore:Support for class-based `config` is deprecated:DeprecationWarning",
    "ignore:pkg_resources is deprecated as an API:DeprecationWarning",
]

[tool.coverage.run]
branch = true
relative_files = true

[tool.ruff]
extend-exclude = [".vscode", ".devcontainer", "swo"]
output-format = "full"
line-length = 100

[tool.ruff.lint]

select = [
    "E", # w errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade,
    "PT", # flake8-pytest-style
    "T10", # flake8-pytest-style
]
ignore = [
    "UP035", # unnecessary open() encoding
    "UP017", # typing updates
    "PT004", # fixture '{name}' does not return anything, add leading underscore
    "PT011", # pytest.raises({exception}) is too broad, set the match parameter or use a more specific exception
    "B008", # do not perform function calls in argument defaults
    "B904", # Within an `except` clause, raise exceptions with `raise ... from err` or `raise ... from None` to distinguish them from errors in exception handling
]

[tool.ruff.lint.isort]
known-third-party = ["swo"]

[tool.ruff.lint.pycodestyle]
max-line-length = 100

[tool.mypy]
warn_no_return = false

[[tool.mypy.overrides]]
module = "adaptive_cards.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "mpt_extension_sdk.*"
ignore_missing_imports = true


[[tool.mypy.overrides]]
module = "django.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pyfiglet.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "gunicorn.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pymsteams.*"
ignore_missing_imports = true
